UAS/UAS.UnitTests/UAS.UnitTests/FrameworkUAD/DataAccess/MarketTest.cs
using System;
using System.Data;
using System.Data.SqlClient;
using System.Data.SqlClient.Fakes;
using System.Diagnostics.CodeAnalysis;
using FrameworkUAD.DataAccess;
using FrameworkUAD.DataAccess.Fakes;
using KMPlatform.Object;
using Microsoft.QualityTools.Testing.Fakes;
using NUnit.Framework;
using Shouldly;
using UAS.UnitTests.Helpers;
using Entity = FrameworkUAD.Entity;

namespace UAS.UnitTests.FrameworkUAD.DataAccess
{
    /// <summary>
    /// Unit tests for <see cref="Market"/>
    /// </summary>
    [TestFixture]
    [ExcludeFromCodeCoverage]
    public class MarketTest
    {
        private const string CommandText = "e_Market_Save";

        private IDisposable _context;
        private SqlCommand _saveCommand;
        private Entity.Market _market;

        [SetUp]
        public void SetUp()
        {
            _context = ShimsContext.Create();
            SetupFakes();
        }

        [TearDown]
        public void TearDown()
        {
            _context.Dispose();
        }

        [Test]
        public void Save_WhenCalled_VerifySqlParameters()
        {
            // Arrange
            _market = typeof(Entity.Market).CreateInstance();

            // Act
            Market.Save(_market, new ClientConnections());

            // Assert
            _market.ShouldSatisfyAllConditions(
                () => _saveCommand.CommandText.ShouldBe(CommandText),
                () => _saveCommand.CommandType.ShouldBe(CommandType.StoredProcedure));

            VerifySqlParameters();
        }

        [Test]
        public void Save_WhenCalledWithNullValues_VerifySqlParameters()
        {
            // Arrange
            _market = typeof(Entity.Market).CreateInstance(true);

            // Act
            Market.Save(_market, new ClientConnections());

            // Assert
            _market.ShouldSatisfyAllConditions(
                () => _saveCommand.CommandText.ShouldBe(CommandText),
                () => _saveCommand.CommandType.ShouldBe(CommandType.StoredProcedure));

            VerifySqlParameters();
        }

        private void VerifySqlParameters()
        {
            _saveCommand.ShouldSatisfyAllConditions(
                () => _saveCommand.Parameters["@MarketID"].Value.ShouldBe(_market.BrandID),
                () => _saveCommand.Parameters["@MarketName"].Value.ShouldBe(_market.BrandID),
                () => _saveCommand.Parameters["@MarketXML"].Value.ShouldBe(_market.BrandID),
                () => _saveCommand.Parameters["@BrandID"].Value.ShouldBe(_market.BrandID),
                () => _saveCommand.Parameters["@DateCreated"].Value.ShouldBe(_market.DateCreated),
                () => _saveCommand.Parameters["@DateUpdated"].Value.ShouldBe((object)_market.DateUpdated ?? DBNull.Value),
                () => _saveCommand.Parameters["@CreatedByUserID"].Value.ShouldBe(_market.CreatedByUserID),
                () => _saveCommand.Parameters["@UpdatedByUserID"].Value.ShouldBe((object)_market.UpdatedByUserID ?? DBNull.Value));
        }

        private void SetupFakes()
        {
            ShimDataFunctions.GetClientSqlConnectionClientConnections = _ => new ShimSqlConnection().Instance;
            ShimDataFunctions.ExecuteScalarSqlCommand = cmd =>
            {
                _saveCommand = cmd;
                return -1;
            };
        }
    }
}